refactor: move ExperimentPostKeyword DTO to `/prompt/dto` for better clarity
