+++ infrastructure/src/main/kotlin/com/dobby/external/gateway/experiment/ExperimentKeywordExtractionGatewayImpl.kt
@@ -3,7 +3,7 @@ package com.dobby.external.gateway.experiment
 import com.dobby.api.dto.request.OpenAiRequest
 import com.dobby.exception.CustomOpenAiCallException
 import com.dobby.external.feign.openAi.OpenAiFeignClient
-import com.dobby.external.prompt.ExperimentPostKeywordDto
+import com.dobby.external.prompt.dto.ExperimentPostKeywordDto
 import com.dobby.external.prompt.ExperimentPostKeywordMapper
 import com.dobby.external.prompt.PromptTemplate
 import com.dobby.external.prompt.PromptTemplateLoader
